+++ WebApiCT/Entities/Models/IngredientRecipe.cs
@@ -10,6 +10,6 @@ namespace CaloriesTracker.Entities.Models
         public Guid RecipeId { get; set; }
         public Recipe Recipe { get; set; }
 
-        public double Grams { get; set; }
+        public float Grams { get; set; }
     }
 }
